requirements.txt
pandas==2.2.3
requests==2.32.5
pydantic==2.12.5
PyYAML==6.0.3
tox==4.49.1
